Added support for IP fragment filtering
The decision to allow a packet is made based on the first IP fragment,
as it contains the L4 header. If the decision is to allow the packet,
an entry is created in the map, and the subsequent IP fragments will
be accepted based on this entry. If there is no entry in the map,
the fragment is blocked.

Limitations:
 * the first fragment must be processed first.
 * in case of a race condition, if a subsequent fragment is
   processed on another CPU simultaneously with the first
   fragment, it may be mistakenly dropped.