Prepare xbuf for handling tc egress traffic
A new field data_end has been added to store the value from
the corresponding field in xdp_md and __sk_buff structures.
It is used to validate memory access bounds when working with
packet data. This allows the same structure to be used for both
xdp_md and __sk_buff contexts.