Add source NAT translation for IP fragments
For the original direction, source IP translation
and TCP/UDP checksum recalculation are applied
to the first fragment, since these headers are
present in the first packet. For all subsequent
fragments, only source IP translation is applied.

For the reply direction, the original IP address
is restored for the first fragment, and a partial
5-tuple is stored in frag_info. Subsequent fragments
use frag_info to reconstruct the full 5-tuple, which
is then used to look up the NAT entry and restore
the original IP address.